# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Locadora de Carros",
        "description": "Documentação da API da Locadora de Carros com autenticação JWT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/v1/cliente": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Clientes cadastrados",
                "operationId": "b673ae9e076460a2f7c8083b65034f17",
                "parameters": [
                    {
                        "name": "filtro",
                        "in": "query",
                        "description": "Filtros dinâmicos. Exemplos de operadores comuns: =, >, >=, <, <=, like, in, between. Use % caso queira pesquisa só o inicio do nome",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "nome:like:J%"
                    },
                    {
                        "name": "atributos",
                        "in": "query",
                        "description": "Quais colunas do recurso 'cliente' retornar. Use vírgula para separar.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "nome,id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os dados do cliente"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Cadastro de clientes",
                "operationId": "58057ca83a1c7140fc79a65e8aa13a3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João Teste"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cliente/{id}": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Detalhes de um cliente",
                "operationId": "a057068d9391a025507495b52447d0b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente a ser pesquisado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Cliente não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Atualização completa de um cliente",
                "description": "Substitui todos os campos do recurso cliente (estilo PUT).",
                "operationId": "162cc3cad204fb75bb882c7dc884c9b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Payload completo para atualização (PUT requer todos os campos obrigatórios do model).",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    },
                    "404": {
                        "description": "Recurso não existe"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Remover um cliente",
                "operationId": "624fdaf82f3eeeb6cae96440d57638c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Cliente não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT no formato: Bearer {seu_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}